CORE-19951 `PersistenceException` handling in uniqueness worker
