--- components/persistence/persistence-service-common/src/main/kotlin/net/corda/persistence/common/PersistenceExceptionCategorizer.kt
@@ -1,8 +1,0 @@
-package net.corda.persistence.common
-
-internal interface PersistenceExceptionCategorizer {
-
-    fun categorize(exception: Exception): PersistenceExceptionType
-}
-
-internal enum class PersistenceExceptionType { FATAL, PLATFORM, TRANSIENT }
