+++ libs/db/db-orm/src/main/kotlin/net/corda/orm/PersistenceExceptionCategorizer.kt
@@ -1,0 +1,8 @@
+package net.corda.orm
+
+interface PersistenceExceptionCategorizer {
+
+    fun categorize(exception: Exception): PersistenceExceptionType
+}
+
+enum class PersistenceExceptionType { FATAL, DATA_RELATED, TRANSIENT, UNCATEGORIZED }
